Update BattleExchangeVariant.cpp
fix.compile error

+++ AI/BattleAI/BattleExchangeVariant.cpp
@@ -290,7 +290,7 @@ ReachabilityInfo getReachabilityWithEnemyBypass(
 				continue;
 
 			auto dmg = damageCache.getOriginalDamage(activeStack, unit, state);
-			auto turnsToKill = unit->getAvailableHealth() / std::max(dmg, 1LL);
+			auto turnsToKill = unit->getAvailableHealth() / std::max(dmg, (int64_t)1);
 
 			vstd::amin(turnsToKill, 100);
